# Only shallow duplicate attributes in `vec_restore()` if they are not `NULL`